Interfaz de algoritmo tokenizador estándar
Por fin obtengo algo que me gusta (se ve coherente): la clase abstracta
define las operaciones de validación; corta la entrada y procesa
(tokeniza) solo el equivalente del identificador de la cuenta. Las
clases concretas (TKR, FFX, BPS) son las que defínen cómo se hace la
operación.

De momento el único implementador es TKR; estoy en proceso de adaptar
FFX y BPS.

+++ implementaciones/redes_feistel/cabeceras/red_feistel.hh
@@ -194,8 +194,8 @@ namespace Implementaciones
     const std::vector<Arreglo<tipo>>& textoEnClaro      /**< Bloque a cifrar. */
   )
   {
-    Arreglo<Arreglo<tipo>> partes = textoEnClaro[0] / Arreglo<int>{
-      static_cast<int>(textoEnClaro[0].obtenerNumeroDeElementos()) / 2};
+    Arreglo<Arreglo<tipo>> partes = textoEnClaro[0] / Arreglo<unsigned int>{
+      textoEnClaro[0].obtenerNumeroDeElementos() / 2};
     Arreglo<tipo> auxiliar (mTamanioDeBloque / 2);
     for (mRondaActual = 0; mRondaActual < mNumeroDeRondas; mRondaActual++)
     {
@@ -229,8 +229,8 @@ namespace Implementaciones
     const std::vector<Arreglo<tipo>>& textoCifrado   /**< Bloque a descifrar. */
   )
   {
-    Arreglo<Arreglo<tipo>> partes = textoCifrado[0] / Arreglo<int>{
-      static_cast<int>(textoCifrado[0].obtenerNumeroDeElementos()) / 2};
+    Arreglo<Arreglo<tipo>> partes = textoCifrado[0] / Arreglo<unsigned int>{
+      textoCifrado[0].obtenerNumeroDeElementos() / 2};
     Arreglo<tipo> auxiliar (mTamanioDeBloque / 2);
     for (mRondaActual = 0; mRondaActual < mNumeroDeRondas; mRondaActual++)
     {
